requirements.txt
streamlit
google-genai
transformers
torch
python-dotenv
